Corrected my typo in training window
Initially GAM was fitted for 2019.12.31-2022.12.31 without rolling window. So, final prediction was for 1 day because N_s - number of samples (days) was defined as 1. Now, N_s is dynamic: changes on training start and end dates will reflected to N_s accordingly. Also, so minor changes for dealing with Series inside gam and forecaster functions.